# lingualab-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Lingualab
  description: Lingualab
  version: 1.0.0
servers:
  - url: 'localhost:8080'
paths:
  '/api/v1/register':
    post:
      tags:
        - auth
      operationId: register
      parameters:
        - in: query
          name: invite
          required: false
          schema:
            type: string
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/RegisterRequestDto"
      responses:
        201:
          description: Ok
  '/api/v1/user/{username}':
    get:
      tags:
        - user
      operationId: getUserByUsername
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
  '/api/v1/user':
    post:
      tags:
        - user
      operationId: createUser
      responses:
        201:
          description: User created
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UserDto'
  '/api/v1/user/{id}':
    get:
      tags:
        - user
      operationId: getUserById

      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    put:
      tags:
        - user
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        202:
          description: Accepted

    delete:
      tags:
        - user
      operationId: deleteUser
      responses:
        202:
          description: Marked as expired
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: expiresAt
          schema:
            type: string
            format: date-time


components:
  schemas:
    RegisterRequestDto:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
        role:
          type: string
        expires_at:
          type: string
          format: date-time
    AuthDto:
      type: object
      properties:
        id:
          type: integer
          format: bigint
        login:
          type: string
        password:
          type: string
        role:
          type: string

    UserDto:
      type: object
      description: информация о пользователе
      properties:
        id:
          description: ID пользователя
          type: integer
          format: int64
        surname:
          description: фамилия
          type: string
        name:
          description: имя
          type: string
        role:
          type: string
        teacher:
          $ref: '#/components/schemas/UserDto'
    PackDto:
      type: object
      description: Набор слов
      properties:
        id:
          description: идентификатор набора
          type: integer
          format: int64
        createdBy:
          $ref: '#/components/schemas/UserDto'
        createdAt:
          description: дата создания
          type: string
          format: date-time
        name:
          description: название набора слов
          type: string

    WordDto:
      type: object
      description: Слово для изучения
      properties:
        id:
          description: идентификатор
          type: integer
          format: int64
        packId:
          description: идентификатор набора
          type: integer
          format: int64
        addedBy:
          description: идентификатор пользователя, добавившего слово
          type: integer
          format: int64
        word:
          type: string
          description: слово
        translation:
          type: string
          description: перевод слова
        transcription:
          type: string
          description: транскрипция слова
          nullable: true
        image:
          type: string
          description: поясняющая картинка
          nullable: true

    Homework:
      type: string
      description: домашняя работа
      properties:
        id:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        teacherId:
          type: integer
          format: int64
        pack:
          $ref: '#/components/schemas/PackDto'
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time